Move file related tasks to file-manager file
